Guia 2 Java POO Extra 02

+++ Guia2_POO_Extras/src/ej02/servicios/ServiciosPuntos.java
@@ -35,9 +35,9 @@ potencia cuadrada de la resta de x2 - x1, mas la
 potencia cuadrada de la resta de y2-y1
         d = √(x2-x1)^2+(y2-y1)^2
 */
-    int f1 = (int)Math.pow ((p.getX2()-p.getX1()),2);
-        int f2 = (int)Math.pow ((p.getY2()-p.getY1()),2);
-        int dis =(int)Math.sqrt(f1+f2);
+        float f1 = (float)Math.pow ((p.getX2()-p.getX1()),2);
+        float f2 = (float)Math.pow ((p.getY2()-p.getY1()),2);
+        float dis =(float)Math.sqrt(f1+f2);
         
         System.out.println("La distancia entre los puntos es: "+dis);
     }
